1.3.D-(1.3 Update 6)

+++ src/main/java/net/gecko/varandeco/block/ice/FragileIceBlock.java
@@ -24,7 +24,7 @@ public class FragileIceBlock extends Block {
     }
 
     @Override
-    public void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, float fallDistance) {
+    public void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, double fallDistance) {
         if (!entity.bypassesSteppingEffects()) {
             if (!entity.getType().isIn(EntityTypeTags.POWDER_SNOW_WALKABLE_MOBS))
                 world.breakBlock(pos, true);
